Don't overwrite `vol->pa_idle_timer`.
The tag of the `pa_update_disp_cb` idle source is stored in
`vol->pa_idle_timer` so that the call can be canceled if the `vol`
struct gets freed before the call happens. That's needed because such a
call would use `vol` after it has been freed, which can cause a segfault
or other unintended behaviour.

However, we can't cancel a call whose tag we have overwritten, so let's
not schedule a call when one is still pending (or, more precisely, when
it hasn't cleared `vol->pa_idle_timer` yet).

The pending call will run its `volumepulse_update_display` call after we
have returned. That call will therefore see the state including the
event that we are currently processing, which is good enough for us.

Fixes: #10